Add a setting for user's base projects folder
This setting (plugins/qcity/base_project_folder) can be set to
a folder to scan for additional base projects within. These
will be added to the base layer combo alongside all projects
which are actually embedded in the plugin.

This permits an administrator to deploy qcity with the setting
pre-filled to a organisation specific folder, eg a network path.

Refs #32