Update installation and update documentation for QCity
Initial documentation for installing and updating QCity. There are a few images missing currently as QCity is not in the QGIS Plugin Repository yet. They will be added later.

Additionally, the support link is to a North Road Consulting e-mail address. @nyalldawson please let me know if this is ok or needs changes.